Handle it better when a managed gRPC server crashes on startup
Summary: This should not be very common, but a user did hit it when the dagster.grpc module wasn't available. Instead of crashing and then waiting 15 seconds for a message that never comes, then timing out, detect right away that the process is no longer alive while waiting for the message.

Test Plan: New Test

Reviewers: alangenfeld, sashank, max

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D5166

+++ python_modules/dagster/dagster_tests/general_tests/grpc_tests/test_persistent.py
@@ -1,8 +1,10 @@
 import os
+import re
 import subprocess
 import uuid
 
 import pytest
+from dagster import seven
 from dagster.core.test_utils import new_cwd
 from dagster.grpc.client import DagsterGrpcClient
 from dagster.grpc.server import wait_for_grpc_server
@@ -39,7 +41,7 @@ def test_ping():
     )
 
     try:
-        wait_for_grpc_server(ipc_output_file)
+        wait_for_grpc_server(process, ipc_output_file)
         assert DagsterGrpcClient(port=port).ping("foobar") == "foobar"
     finally:
         process.terminate()
@@ -68,11 +70,12 @@ def test_load_with_error(capfd):
 
     try:
         with pytest.raises(DagsterIPCProtocolError):
-            wait_for_grpc_server(ipc_output_file)
+            wait_for_grpc_server(process, ipc_output_file)
         _, err = capfd.readouterr()
         assert "No module named" in err
     finally:
-        process.terminate()
+        if process.poll() is None:
+            process.terminate()
 
 
 def test_load_with_non_existant_file(capfd):
@@ -114,7 +117,7 @@ def test_load_with_empty_working_directory(capfd):
         )
 
         try:
-            wait_for_grpc_server(ipc_output_file)
+            wait_for_grpc_server(process, ipc_output_file)
             assert DagsterGrpcClient(port=port).ping("foobar") == "foobar"
         finally:
             process.terminate()
@@ -140,10 +143,45 @@ def test_load_with_empty_working_directory(capfd):
         )
         try:
             with pytest.raises(DagsterIPCProtocolError):
-                wait_for_grpc_server(ipc_output_file)
+                wait_for_grpc_server(process, ipc_output_file)
             _, err = capfd.readouterr()
             assert "No module named" in err
         finally:
+            if process.poll() is None:
+                process.terminate()
+
+
+@pytest.mark.skipif(seven.IS_WINDOWS, reason="Crashes in subprocesses crash test runs on Windows")
+def test_crash_during_load():
+    port = find_free_port()
+    python_file = file_relative_path(__file__, "crashy_grpc_repo.py")
+
+    ipc_output_file = _get_ipc_output_file()
+
+    process = subprocess.Popen(
+        [
+            "dagster",
+            "api",
+            "grpc",
+            "--port",
+            str(port),
+            "--python-file",
+            python_file,
+            "--ipc-output-file",
+            ipc_output_file,
+        ],
+        stdout=subprocess.PIPE,
+    )
+    try:
+
+        with pytest.raises(
+            Exception,
+            match=re.escape("Process exited with return code 123 while waiting for events"),
+        ):
+            wait_for_grpc_server(process, ipc_output_file)
+
+    finally:
+        if process.poll() is None:
             process.terminate()
 
 
@@ -170,7 +208,7 @@ def test_lazy_load_with_error():
     )
 
     try:
-        wait_for_grpc_server(ipc_output_file)
+        wait_for_grpc_server(process, ipc_output_file)
         list_repositories_response = DagsterGrpcClient(port=port).list_repositories()
         assert isinstance(list_repositories_response, SerializableErrorInfo)
         assert "No module named" in list_repositories_response.message
@@ -200,7 +238,7 @@ def test_streaming():
     )
 
     try:
-        wait_for_grpc_server(ipc_output_file)
+        wait_for_grpc_server(process, ipc_output_file)
         api_client = DagsterGrpcClient(port=port)
         results = [result for result in api_client.streaming_ping(sequence_length=10, echo="foo")]
         assert len(results) == 10
